etl/requirements.txt
google-cloud-storage
google-cloud-bigquery
requests
functions-framework
